requirements.txt
spacy
python-telegram-bot
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
pandas==2.1.4